test: combine all unit tests into a single file
