+++ vite.config.js
@@ -58,7 +58,7 @@ export default defineConfig(({ mode }) => {
       },
     },
     test: {
-      include: ['**/tests/unit/**/*.spec.{j,t}s?(x)'],
+      include: ['**/tests/unit/**/test.spec.{j,t}s?(x)'],
       threads: false,
       globals: true,
       environment: 'node',
